encapsulate the publish dm method
